move threading helpers out of compat module
