+++ tests/core/filtering/test_filter_against_transaction_logs.py
@@ -2,7 +2,7 @@ import pytest
 import random
 from flaky import flaky
 
-from web3.utils.compat import (
+from web3.utils.threads import (
     Timeout,
 )
