+++ tests/core/mining-module/test_miner_setExtra.py
@@ -6,7 +6,7 @@ from eth_utils import (
     decode_hex,
 )
 
-from web3.utils.compat import (
+from web3.utils.threads import (
     Timeout,
 )
