+++ tests/core/mining-module/test_miner_stop.py
@@ -2,7 +2,7 @@ import random
 
 from flaky import flaky
 
-from web3.utils.compat import (
+from web3.utils.threads import (
     Timeout,
 )
