+++ tests/core/txpool-module/test_txpool_content.py
@@ -1,6 +1,6 @@
 import random
 
-from web3.utils.compat import (
+from web3.utils.threads import (
     Timeout,
 )
