+++ web3/providers/tester.py
@@ -17,9 +17,6 @@ from web3.middleware import (
     construct_exception_handler_middleware,
 )
 
-from web3.utils.compat import (
-    spawn,
-)
 from web3.utils.formatters import (
     hex_to_integer,
     apply_formatter_if,
@@ -27,6 +24,9 @@ from web3.utils.formatters import (
     apply_formatters_to_dict,
     static_return,
     static_result,
+)
+from web3.utils.threads import (
+    spawn,
 )
 
 from .base import BaseProvider  # noqa: E402
