+++ web3/utils/compat/__init__.py
@@ -14,10 +14,6 @@ THREADING_BACKEND = get_threading_backend()
 
 
 if THREADING_BACKEND == 'stdlib':
-    from .compat_stdlib import (  # noqa: F401
-        Timeout,
-        spawn,
-    )
     from .compat_requests import (  # noqa: F401
         make_post_request,
     )
